moved firebase types to common
